app-tier-python/requirements.txt
flask==3.1.0
flask_cors==5.0.0
mysql-connector-python==9.1.0